+++ src/common/matrix.h
@@ -14,8 +14,8 @@ class Matrix {
   explicit Matrix(int32_t row_count, int32_t col_count) :
     row_count_(row_count),
     col_count_(col_count),
-    elements_(std::vector<float>(row_count_ * col_count_)) {};
-  explicit Matrix(int32_t row_count, int32_t col_count, std::vector<float> elements) :
+    elements_(std::vector<double>(row_count_ * col_count_)) {};
+  explicit Matrix(int32_t row_count, int32_t col_count, std::vector<double> elements) :
     row_count_(row_count),
     col_count_(col_count),
     elements_(std::move(elements)) {
@@ -25,25 +25,25 @@ class Matrix {
   static Matrix Random(int32_t row_count, int32_t col_count);
 
   Matrix Transpose() const;
-  Matrix Map(std::function<float(float)> closure) const;
-  Matrix Merge(const Matrix& other, std::function<float(float, float)> closure) const;
+  Matrix Map(std::function<double(double)> closure) const;
+  Matrix Merge(const Matrix& other, std::function<double(double, double)> closure) const;
   Matrix HadamardMult(const Matrix& other) const;
   Matrix operator+(const Matrix& other) const;
-  Matrix operator*(float scalar) const;
+  Matrix operator*(double scalar) const;
   Matrix operator*(const Matrix& other) const;
   bool operator==(const Matrix& other) const;
 
   int32_t Classify() const;
   int32_t RowCount() const;
   int32_t ColCount() const;
-  float ElementAt(int32_t r, int32_t c) const;
-  float& MutableElementAt(int32_t r, int32_t c);
+  double ElementAt(int32_t r, int32_t c) const;
+  double& MutableElementAt(int32_t r, int32_t c);
   std::string DebugString() const;
 
  private:
   int32_t row_count_;
   int32_t col_count_;
-  std::vector<float> elements_;
+  std::vector<double> elements_;
 };
 
 #endif
